Using pydantic to save and load json safely.
renamed shmup.py to __main__.py so that the whole project can be treated as a package that can be ran.

moved image loading functionality to load.py

fixed text aligning issue by creating wrap_rects that have absolute values based on the actual size of the text.

now using a dataclass for the system data.

renamed config to settings.

adding more logging